% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_google.R
\name{db_find}
\alias{db_find}
\title{Which row is the user on in the DB}
\usage{
db_find(x, out_type = logical(), col = c("email", "user")[1])
}
\arguments{
\item{out_type}{\code{obj} of type corresponding to output}

\item{email}{\code{chr} User email address}
}
\value{
\code{obj} of type similar to `out_type`
}
\description{
Which row is the user on in the DB
}
