% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_google.R
\name{db_row_update}
\alias{db_row_update}
\title{Update a row in database}
\usage{
db_row_update(.row, range = NULL, col = "user")
}
\arguments{
\item{.row}{\code{tbl} Single row of a tbl}

\item{range}{\code{cell_limits} to update}
}
\value{
\code{msg}
}
\description{
Update a row in database
}
