% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{get_admins}
\alias{get_admins}
\title{Load the table of admin names}
\usage{
get_admins(filepath = fs::path("data", "admins", ext = "feather"))
}
\arguments{
\item{filepath}{\code{(character)} path to admin table file}
}
\value{
\code{(tibble)}
}
\description{
Load the table of admin names
}
