% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{is_admin}
\alias{is_admin}
\title{Checks if a first name / last name pair is an admin}
\usage{
is_admin(first_name, last_name)
}
\arguments{
\item{first_name}{\code{(character)}}

\item{last_name}{\code{(character)}}
}
\value{
\code{(logical)}
}
\description{
Sets global `user_profile` state `admin = TRUE`
}
