% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%||\%}
\alias{\%|\%}
\alias{\%|0|\%}
\alias{\%|try|\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang:op-na-default]{\%|\%}}, \code{\link[rlang:op-null-default]{\%||\%}}}

  \item{UU}{\code{\link[UU:grapes-or-0-or-grapes]{\%|0|\%}}, \code{\link[UU:tick-grapes-or-try-or-grapes-tick]{\%|try|\%}}}
}}

