% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_time_picker.R
\name{time_handler}
\alias{time_handler}
\title{Convert between tibble and character representations of time}
\usage{
time_handler(x)
}
\arguments{
\item{x}{\code{tbl/chr} tbl with columns day, begin, end. chr in the format `"Mon_4:0_4:0"`}
}
\value{
\code{tbl/chr}
}
\description{
Convert between tibble and character representations of time
}
