% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{user_times}
\alias{user_times}
\title{Fetch user times or return a default time}
\usage{
user_times(user)
}
\arguments{
\item{user}{\code{chr} username}
}
\value{
\code{tbl} with rows of user times
}
\description{
Fetch user times or return a default time
}
